Make keyed mutation queues deterministic
